+++ kolibri/core/exams/api_urls.py
@@ -2,7 +2,7 @@ from django.urls import include
 from django.urls import re_path
 from rest_framework import routers
 
-from .api import ExamViewset
+from .viewsets.exam import ExamViewset
 
 router = routers.SimpleRouter()
 router.register(r"exam", ExamViewset, basename="exam")
